\name{Boots-package}
\alias{Boots-package}
\alias{Boots}
\docType{package}
\title{
  Predicts Disposition of U.S. Military Personnel by Country and Branch
}
\description{
 This package utilizes R packages to forecast the number of U.S. Military personnel by Country in either a disaggregated fashion (by branch) or an aggregated fashion (All Services).  The results are plotted using ggplot2.
}
\details{
  Run the app using Shiny to view predictions.
}
\author{Matthew Small <matthew.small@afit.edu>
}
\references{
  See Hyndman Automatic Forecasting (2008)
}
\keyword{ forecast, military, personnel }
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
