% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_my_app.R
\name{run_my_app}
\alias{run_my_app}
\title{Runs the shiny app for my package}
\usage{
run_my_app(app_name, ...)
}
\arguments{
\item{app_name}{Character string for a directory in this package}

\item{...}{Additional options passed to shinyAppDir}
}
\description{
Runs as shiny app
}
\examples{
\dontrun{ run_my_app("Boots")}
}
